requirements.txt
pandas
streamlit
beautifulsoup4
requests
openpyxl
plotly
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
icalendar
